Use time.perf_counter if possible to measure time periods
With this commit we will use `time.perf_counter()` if available
to measure time periods. On older Python versions we will
fall back to `time.time()`.

Previously, `datetime.datetime.now()` has been used which
resulted in lots of system calls opening the system timezone
file. With this implementation we'll avoid that.

Note that callsites have not been changed yet and we only
adapted `ExpirationTimer` / `ExpiryTime` to use the new
implementation.